==> training/1.prepare_data.ipynb <==
import pandas as pd
import json
import pickle

with open(r"C:\Users\deviy\Desktop\reco\Foodfinder-Recommendation-System\reco_data\yelp_academic_dataset_business.json", encoding="utf8") as business_file:
    business_data = [json.loads(line) for line in business_file]
business_df = pd.DataFrame(business_data)

# Load user data
with open(r"C:\Users\deviy\Desktop\reco\Foodfinder-Recommendation-System\reco_data\yelp_academic_dataset_user.json", encoding="utf8") as user_file:
    user_data = [json.loads(line) for line in user_file]
user_df = pd.DataFrame(user_data)

# Load review data
with open(r"C:\Users\deviy\Desktop\reco\Foodfinder-Recommendation-System\reco_data\yelp_academic_dataset_review.json", encoding="utf8") as review_file:
    review_data = [json.loads(line) for line in review_file]
review_df = pd.DataFrame(review_data)

**Filter Loc&Categ of Business**

nash_business_df=business_df.loc[business_df['city']=='Nashville']

nash_business_df.categories.value_counts()

len(nash_business_df)#.categories.value_counts()

del business_df

# Filter Nashville restaurants and food-related businesses
nashville_restaurants_df = nashville_business_df[
    nashville_business_df['categories'].str.contains('Restaurants|Food') & 
    nashville_business_df['categories'].notna()
]

with open('nash_restaurants.pkl','wb') as f:          
    pickle.dump(nashville_restaurants_df,f)

# Get list of Nashville restaurant business IDs
nashville_restaurant_ids = nashville_restaurants_df['business_id'].tolist()

nashville_restaurant_ids 

with open('nash_resto_ids.json', 'w') as fp:
    json.dump(nashville_restaurant_ids, fp)

# Get list of open Nashville restaurant business IDs
nashville_open_restaurant_ids = nashville_restaurants_df[nashville_restaurants_df['is_open'] != 0]['business_id'].tolist()

with open('nash_open_ids.json', 'w') as fp:
    json.dump(nashville_open_restaurant_ids, fp)

**Filter important users & Reviews**

important_users_df = user_df[(user_df['review_count'] >= 10) | (user_df['elite'].apply(lambda x: len(str(x)) > 0))]

# Get list of important user IDs
important_user_ids = important_users_df['user_id'].tolist()

del user_df

# Filter reviews for important users and Nashville restaurants
nashville_reviews_df = review_df[
    review_df['user_id'].isin(important_user_ids) &
    review_df['business_id'].isin(nashville_restaurant_ids)
]


# Calculate average stars for each Nashville restaurant
nashville_stars_df = nashville_reviews_df[nashville_reviews_df['business_id'].isin(nashville_restaurants_df['business_id'])].groupby('business_id')['stars'].mean().reset_index()

with open('nash_stars.pkl','wb') as f:          
    pickle.dump(nashville_stars_df, f)

# Filter important users who have reviewed Nashville restaurants
nashville_important_users_df = important_users_df[important_users_df['user_id'].isin(nashville_reviews_df['user_id'])]

with open('nash_reviews.pkl','wb') as f:          
    pickle.dump(nashville_reviews_df, f)

with open('nash_user.pkl','wb') as f:         
    pickle.dump(nashville_important_users_df, f)

nashville_important_users_df.shape
