==> training/2.collaborative_filtering.ipynb <==
import warnings
warnings.filterwarnings('ignore')
import os
import random
import json
import numpy as np
import pandas as pd
import pickle
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

sns.set()

!pip install surprise

from surprise import SVD, NMF, NormalPredictor,KNNBasic
from surprise import accuracy, Reader, Dataset, dump 
from surprise.model_selection import cross_validate, GridSearchCV, train_test_split

# Load review data
reviews_df = pd.read_pickle("nash_reviews.pkl")

unique_user_ids = reviews_df['user_id'].unique()
resto = reviews_df['business_id'].unique()

user_map = {unique_user_ids[i]: i + 1 for i in range(len(unique_user_ids))}
resto_map = {resto[i]: i + 1 for i in range(len(resto))}

## Preparing data for the surprise library

reader = Reader(rating_scale=(1, 5))
data = Dataset.load_from_df(reviews_df[['user_id', 'business_id', 'stars']], reader)

# Evaluating different algorithms
performance = []
for algo in [SVD(), NMF(), NormalPredictor()]:
    cv_results = cross_validate(algo, data, measures=['RMSE'], cv=3, verbose=False)
    results_df = pd.DataFrame.from_dict(cv_results).mean(axis=0)
    results_df = results_df._append(pd.Series([str(algo).split(' ')[0].split('.')[-1]], index=['Algorithm']))
    performance.append(results_df)

pd.DataFrame(performance).set_index('Algorithm')

### Pick SVD, as it seems to perform better

# Grid search for best parameters
param_grid = {'n_factors': [30, 35], 'n_epochs': [20, 25], 'lr_all': [0.001, 0.003], 'reg_all': [0.08, 0.1]}
grid_search = GridSearchCV(SVD, param_grid, measures=['rmse', 'mae'], cv=3)
grid_search.fit(data)


best_algo = grid_search.best_estimator['rmse']
print(grid_search.best_score['rmse'])
print(grid_search.best_params['rmse'])

# Training on the entire dataset
trainset = data.build_full_trainset()
best_algo.fit(trainset)

# Saving the trained model
file_name = os.path.expanduser('best_algo_dump')
dump.dump(file_name, algo=best_algo)

# Loading the saved model
_, loaded_best_algo = dump.load(file_name)

# Try to predict recommendation for each user
user_list = {i:j for i,j in enumerate(reviews_df['user_id'].unique())}

# Function to get top N recommendations
from collections import defaultdict
def get_top_n(predictions, n=10):
    """Return the top-N recommendation for each user from a set of predictions."""
    top_n = defaultdict(list)
    for uid, iid, true_r, est, _ in predictions:
        top_n[uid].append((iid, est))
    for uid, user_ratings in top_n.items():
        user_ratings.sort(key=lambda x: x[1], reverse=True)
        top_n[uid] = user_ratings[:n]
    return top_n

len(user_list.values())

batch_size = 10000
total_users = len(user_list)

for start_idx in range(0, total_users, batch_size):
    print(start_idx)
    batch_users = list(user_list.keys())[start_idx:start_idx + batch_size]
    anti_testset = []
    fill_value = trainset.global_mean
    for uid in batch_users:
        user_items = [item for (item, _) in trainset.ur[uid]]
        for iid in trainset.all_items():
            if iid not in user_items:
                anti_testset.append((trainset.to_raw_uid(uid), trainset.to_raw_iid(iid), fill_value))
    
    predictions = loaded_best_algo.test(anti_testset)
    top_n_recommendations = get_top_n(predictions, n=5)
    
    with open('rec_test_{}.json'.format(start_idx // batch_size + 1), 'w') as fp:
        json.dump(top_n_recommendations, fp)
    
    print(f"Processing batch {start_idx // batch_size + 1}: {batch_users[:5]} ... {batch_users[-5:]}")

The top predictions for a select few have been saved in rec.json

nash_restaurants = pd.read_pickle("nash_restaurants.pkl")

nash_restaurants.head()

list(nash_restaurants.postal_code.unique())

zipc=pd.read_csv('us_zip_code.csv')

zipc['postal code'] = zipc['postal code'].astype(str)

len(zipc[zipc['postal code'].isin(nash_restaurants.postal_code.unique())])

len(nash_restaurants.postal_code.unique())


zipc[zipc['postal code'].isin(nash_restaurants.postal_code.unique())][['postal code','geopoint', 'place name']]

zipc['geopoint'] = zipc.apply(lambda row: [row['latitude'], row['longitude']], axis=1)


zipc[zipc['postal code'].isin(nash_restaurants.postal_code.unique())][['postal code','geopoint', 'place name']].to_csv('nash_zip.csv',index=False)
