==> training/4.content-reco&cold start.ipynb <==
# !pip install matplotlib
# !pip install seaborn

# !pip install numpy==1.23
import warnings
warnings.filterwarnings('ignore')
import os
import random
import json
import numpy as np
import pandas as pd
import pickle
import matplotlib.pyplot as plt
import seaborn as sns
import datetime
%matplotlib inline

from scipy import stats
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import linear_kernel
import folium
from folium import plugins
from folium.plugins import HeatMap

with open('nash_contents.pkl','rb') as f:
    nash_contents = pickle.load(f)

nash_contents.head()

# Reading the json file
user_ids = {}

# Loop through each file from 1 to 8
for i in range(1, 9):
    filename = f"rec_test_{i}.json"
    try:
        with open(filename, 'r') as fp:
            # Load data from the file
            data = json.load(fp)
            # Merge the loaded data into the user_ids dictionary
            user_ids.update(data)
    except FileNotFoundError:
        # If file not found, continue to the next iteration
        continue
zips=pd.read_csv('nash_zip.csv',sep=',', quotechar='"', converters={'geopoint': eval})

zip_codes = nash_contents['postal_code'].unique().tolist()

zips['postal code']=zips['postal code'].astype(str)

zip_site = list(zips['postal code'])

[item for item in zip_codes if item not in zip_site]

locations = {str(item['postal code']):item['geopoint'] for index, item in zips.iterrows() if str(item['postal code']) in zip_codes}

[(item,item) for item in list(map(str,sorted([int(item) for item in sorted(locations.keys())])))]

locations.keys()

topic_mapping = {
    0: 'Pizza and Italian Food',
    1: 'Mexican Food',
    2: 'Breakfast and Coffee',
    3: 'Fried Chicken and Fast Food',
    4: 'Casual Dining and Bars',
    5: 'Barbecue'
}

def map_topic(num):
    return topic_mapping[num]

nash_contents['Topics'] = nash_contents['Dominant Ids'].apply(lambda x: list(map(map_topic,x)))

# approximate radius of earth in km
def distance(coord1,coord2):
    R = 6373.0
    lat1,lon1 = coord1
    lat2,lon2 = coord2
    lat1,lon1 = np.radians(lat1), np.radians(lon1)
    lat2,lon2 = np.radians(lat2), np.radians(lon2)
    dlon = lon2 - lon1
    dlat = lat2 - lat1

    a = np.sin(dlat/2)**2 + np.cos(lat1)*np.cos(lat2)*np.sin(dlon/2)**2
    c = 2*np.arctan2(np.sqrt(a), np.sqrt(1 - a))

    dist = R*c/1.6
    return dist

coord1 = (36.2296756, -115.0122287)
coord2 = (36.2006374, -115.2251681)
print(f"The distance between the two is about {int(distance(coord1,coord2)):2} miles.")

zip_std = '37210'
for zip in ['32709', '37203', '37189', '37204']:
    if (distance(locations[zip_std],locations[zip])) < 20.0:
        print(f'The distance from {zip_std} to {zip} {locations[zip]} is {(distance(locations[zip_std],locations[zip])):9.3f} miles')

nash_contents['Features'] = nash_contents['Features'].apply(lambda x: [item for item in x if item not in ['Restaurants','Food']])

nash_contents['Combined Features'] = (nash_contents['Features']).apply(lambda x: ' '.join(x))

nash_contents.head()

with open('nash_restaurants.pkl', 'wb') as fp:
    pickle.dump(nash_contents.drop(columns=['stars_x']).rename(columns={'stars_y':'stars'}), fp)

nash_contents.head()

topic = 'Pizza and Italian Food'
get_restaurants_by_topic(nash_contents,categories_list=[topic])[:2]


preference = 'Chicken Shop'
get_restaurants_by_preference(nash_contents,categories_list=[preference])[:2]

preference1 = ['Chinese']
preference2 = ['Indian']
topic = ['Breakfast and Coffee']
store_rest = []

rest_pref1 = get_restaurants_by_preference(nash_contents,preference1)
rest_pref2 = get_restaurants_by_preference(nash_contents,preference2)
pref_store = [rest_pref1,rest_pref2]

if preference1 != preference2:
    for i in range(2):
        store_rest.extend(pref_store[i][:10])
else:
    store_rest.extend(pref_store[0][:20])
store_rest.extend(get_restaurants_by_topic(nash_contents,topic)[:10])

zip_code = '37232'
proximity = 20.0
for item in store_rest:
    dist = distance(locations[zip_code],locations[item['Zip Code']])    
    if dist < proximity:        
        print(f'The distance from {zip_code} is {dist:9.3f} miles')

[item for item in store_rest if distance(locations[zip_code],locations[item['Zip Code']]) < proximity][:2]

nash_coordinates = (36.174465, - -86.767960)
folium_map = folium.Map(location=nash_coordinates,width=1000,height=800,zoom_start=11)

for restaurant in store_rest:
    coordinates = restaurant['Latitude'],restaurant['Longitude']
    folium.Marker(coordinates,popup=(restaurant['Name'],restaurant['Category'])).add_to(folium_map)

data = [(restaurant['Latitude'],restaurant['Longitude']) for restaurant in store_rest]
HeatMap(data).add_to(folium.FeatureGroup(name='Heat Map').add_to(folium_map))


folium_map

# with open('restaurants.pkl','rb') as f:
#     restaurants = pickle.load(f)

tfidf = TfidfVectorizer()

tfidf_matrix = tfidf.fit_transform(nash_contents['Combined Features'])

cosine_sim = linear_kernel(tfidf_matrix, tfidf_matrix)

cosine_sim[0]

nash_contents.index

nash_contents['business_id']

businesses = nash_contents['business_id']
indices = pd.Series(businesses.index,index = nash_contents['business_id'])

def content_recommendations(business):
    idx = indices[business]
    sim_scores = list(enumerate(cosine_sim[idx]))
    sim_scores = sorted(sim_scores, key=lambda x: x[1], reverse=True)
    sim_scores = sim_scores[1:21]
    biz_indices = [i[0] for i in sim_scores]
    return businesses.iloc[biz_indices].values.tolist()

content_recommendations('bBDDEgkFA1Otx9Lfe7BZUQ')

nash_reviews = pd.read_pickle('nash_reviews.pkl')

nash_reviews['date'] = nash_reviews['date'].apply(lambda x: datetime.datetime.strptime(x, "%Y-%m-%d %H:%M:%S"))

nash_reviews_open = nash_reviews[nash_reviews['business_id'].isin(nash_contents['business_id'])]

nash_reviews_open.head()

last_restaurants = nash_reviews_open[['user_id','business_id','date']].merge(nash_reviews_open.groupby('user_id')['date'].max(),on=['user_id','date'])

usr_biz = []

for usr in user_ids.keys():
    user_data = last_restaurants[last_restaurants['user_id'] == usr]['business_id']
    if not user_data.empty:
        usr_biz.append((usr, user_data.iloc[0]))

store_restaurants_content = {item[0]:content_recommendations(item[1]) for item in usr_biz}

with open('rec_content.json', 'w') as fp:
    json.dump(store_restaurants_content, fp)

test_user = '2XKlUjHGk9z-2ckFMcBEHw'

store_restaurants_content[test_user]

def get_restaurants_by_usr_id(df,usr_id):
    
    store_restaurants = []
    
    df_colab = nash_contents[nash_contents['business_id'].isin([item[0] for item in user_ids[usr_id]])]
    df_content = nash_contents[nash_contents['business_id'].isin([item for item in store_restaurants_content[usr_id]])]
    df_rest = pd.concat([df_colab,df_content])
     
    for index,row in df_rest.iterrows():
        restaurant_details = {}
        restaurant_details['Name'] = row['name']
        restaurant_details['Address'] = row['address']
        restaurant_details['Zip Code'] = row['postal_code']
        restaurant_details['Latitude'] = row['latitude']
        restaurant_details['Longitude'] = row['longitude']
        
        restaurant_details['Category'] = row['sep_categories']
        restaurant_details['Topics'] = row['Topics']
        restaurant_details['Stars'] = row['stars_y']
#         restaurant_details['additional_details'] = row['attributes']
        store_restaurants.append(restaurant_details)
    return sorted(store_restaurants,key = lambda x: x['Stars'],reverse=True)

store = get_restaurants_by_usr_id(nash_contents,test_user)
[item for item in store if distance(locations[zip_code],locations[item['Zip Code']]) < proximity]

### Cold Start

# Calculate the mean star rating for each restaurant
restaurant_means = nash_contents.groupby('business_id')['stars_y'].mean().reset_index()

# Merge mean star ratings back to the original DataFrame
nash_contents_with_means = pd.merge(nash_contents, restaurant_means, on='business_id', suffixes=('', '_mean'))

# Group restaurants by postal code and find the top 10 highest rated restaurants in each group
top_10_by_postal_code = nash_contents_with_means.groupby('postal_code').apply(lambda x: x.nlargest(10, 'stars_y_mean'))

# Reset index to flatten the DataFrame
top_10_by_postal_code = top_10_by_postal_code.reset_index(drop=True)

top_10_by_postal_code.loc[top_10_by_postal_code['postal_code']!=''][['name','address','postal_code','latitude','longitude','sep_categories','Topics','stars_y']].rename(columns={'stars_y':'stars'}).to_csv('cold_start.csv',index=False)
